+++ core/vil1/vil1_byte_swap.cxx
@@ -8,14 +8,15 @@
 
 #include "vil1_byte_swap.h"
 #include <vcl_cassert.h>
+#include <vcl_cstddef.h>
 
 // this is not efficient
 void vil1_byte_swap(void *b_, void *e_) {
   char *b = static_cast<char*>(b_);
   char *e = static_cast<char*>(e_);
   assert(b < e);
-  unsigned long n = e-b;
-  for (unsigned long i=0; i<n/2; ++i) {
+  ptrdiff_t n = e-b;
+  for (ptrdiff_t i=0; i<n/2; ++i) {
     char tmp = b[i];
     b[i] = b[n-1-i];
     b[n-1-i] = tmp;
